Сделал задание 4
