+++ index.html
@@ -8,7 +8,6 @@
   <title>Кексобукинг</title>
 </head>
 <body>
-  <script src ="js/main.js" defer></script>
 
   <main>
     <div class="promo">
@@ -229,5 +228,7 @@
     </div>
   </template>
 
+
+  <script  type="module" src ="js/main.js"></script>
 </body>
 </html>
