+++ js/random-array-element.js
@@ -1,0 +1,7 @@
+import {getRandomNumber} from './random-number.js';
+
+const getRandomArrayElement = (elements) => {
+  return elements[getRandomNumber(0, elements.length - 1)];
+}
+
+export {getRandomArrayElement};
